+++ tests/panels/test_cache.py
@@ -45,6 +45,42 @@ class CachePanelTestCase(BaseTestCase):
         self.assertEqual(self.panel.hits, 4)
         self.assertEqual(self.panel.misses, 2)
 
+    def test_get_default_counts_as_miss(self):
+        """
+        Using get that fetches a cached value that's the same as the default
+        results in a miss despite it being a hit.
+        """
+        cache.cache.clear()
+
+        # cached None with default None is ambiguous
+        cache.cache.set("foo", None)
+        self.assertIsNone(cache.cache.get("foo"))
+        self.assertIsNone(cache.cache.get("foo"))
+        self.assertEqual(self.panel.misses, 2)
+        self.assertEqual(self.panel.hits, 0)
+
+        # cached value equal to default is ambiguous
+        cache.cache.set("bar", "baz")
+        self.assertEqual(cache.cache.get("bar", "baz"), "baz")
+        self.assertEqual(cache.cache.get("bar", "baz"), "baz")
+        self.assertEqual(self.panel.misses, 4)
+        self.assertEqual(self.panel.hits, 0)
+
+    def test_get_with_args(self):
+        """
+        Test get to calculate the default in a variety of ways.
+        """
+        cache.cache.clear()
+        self.assertIsNone(cache.cache.get("foo"))
+        self.assertEqual(self.panel.misses, 1)
+        self.assertEqual(cache.cache.get("foo", "bar"), "bar")
+        self.assertEqual(self.panel.misses, 2)
+        self.assertEqual(cache.cache.get("foo", default="bar"), "bar")
+        self.assertEqual(self.panel.misses, 3)
+        self.assertEqual(cache.cache.get(key="foo", default="bar"), "bar")
+        self.assertEqual(self.panel.misses, 4)
+        self.assertEqual(self.panel.hits, 0)
+
     def test_get_or_set_value(self):
         cache.cache.get_or_set("baz", "val")
         self.assertEqual(cache.cache.get("baz"), "val")
@@ -113,6 +149,17 @@ class CachePanelTestCase(BaseTestCase):
                 "decr_version": 0,
             },
         )
+
+    def test_get_or_set_none_counts_as_miss(self):
+        """
+        Using get_or_set to set a None value is counted as a miss even
+        though it was a hit on the second try.
+        """
+        cache.cache.clear()
+        self.assertIsNone(cache.cache.get_or_set("foo", None))
+        self.assertIsNone(cache.cache.get_or_set("foo", None))
+        self.assertEqual(self.panel.misses, 2)
+        self.assertEqual(self.panel.hits, 0)
 
     def test_insert_content(self):
         """